requirements.txt
streamlit
pandas
google-generativeai
openpyxl
python-dotenv